requirements.txt
google-generativeai
chonkie
neo4j
python-dotenv
tiktoken
